Move Language specific tests to the language layer
