+++ src/Microsoft.VisualStudio.ProjectSystem.Managed.VS.UnitTests/Mocks/IRoslynServicesFactory.cs
@@ -2,26 +2,24 @@
 
 using Moq;
 using Microsoft.CodeAnalysis;
-using System.Threading.Tasks;
+using Microsoft.VisualStudio.ProjectSystem.LanguageServices;
 
 namespace Microsoft.VisualStudio.ProjectSystem.VS
 {
     internal static class IRoslynServicesFactory
     {
-        public static IRoslynServices Create()
-        {
-           return Mock.Of<IRoslynServices>();
-        }
-
-        public static IRoslynServices Implement(Solution solution, bool changesApplied)
+        public static IRoslynServices Implement(ISyntaxFactsService syntaxFactsService)
         {
             var mock = new Mock<IRoslynServices>();
-           
-            mock.Setup(h => h.RenameSymbolAsync(It.IsAny<Solution>(), It.IsAny<ISymbol>(), It.IsAny<string>()))
-                .Returns(Task.FromResult(solution));
 
-            mock.Setup(h => h.ApplyChangesToSolution(It.IsAny<Workspace>(), It.IsAny<Solution>()))
-                .Returns(changesApplied);
+            mock.Setup(h => h.IsValidIdentifier(It.IsAny<string>()))
+                .Returns<string>(name => syntaxFactsService.IsValidIdentifier(name));
+
+            mock.Setup(h => h.IsModuleDeclaration(It.IsAny<SyntaxNode>()))
+                .Returns<SyntaxNode>(node => syntaxFactsService.IsModuleDeclaration(node));
+
+            mock.Setup(h => h.GetModuleName(It.IsAny<SyntaxNode>()))
+                .Returns<SyntaxNode>(node => syntaxFactsService.GetModuleName(node));
 
             return mock.Object;
         }
