+++ src/Microsoft.VisualStudio.ProjectSystem.Managed.VS.UnitTests/ProjectSystem/VS/Rename/SimpleRenamerTestsBase.cs
@@ -1,0 +1,59 @@
+﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
+
+using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.Text;
+
+namespace Microsoft.VisualStudio.ProjectSystem.VS.Rename
+{
+    public abstract class SimpleRenamerTestsBase
+    {
+        protected abstract string TempFileName { get; }
+        protected abstract string DefaultFile { get; }
+
+        protected abstract string ProjectFileExtension { get; }
+
+        protected Solution DummySolution(
+            string tempFileName = null,
+            string defaultFile = null,
+            string language = LanguageNames.CSharp)
+        {
+            using (var ws = new AdhocWorkspace())
+            {
+                return ws.AddSolution(
+                    InitializeWorkspace(
+                        ProjectId.CreateNewId(),
+                        tempFileName ?? TempFileName,
+                        defaultFile ?? DefaultFile,
+                        language));
+            }
+        }
+
+        protected SolutionInfo InitializeWorkspace(ProjectId projectId, string fileName, string code, string language)
+        {
+            var solutionId = SolutionId.CreateNewId();
+            var documentId = DocumentId.CreateNewId(projectId);
+            return SolutionInfo.Create(
+                solutionId,
+                VersionStamp.Create(),
+                projects: new ProjectInfo[]
+                {
+                    ProjectInfo.Create(
+                        id: projectId,
+                        version: VersionStamp.Create(),
+                        name: "Project1",
+                        assemblyName: "Project1",
+                        filePath: $@"C:\project1.{ProjectFileExtension}",
+                        language: language,
+                        documents: new DocumentInfo[]
+                        {
+                                DocumentInfo.Create(
+                                documentId,
+                                fileName,
+                                loader: TextLoader.From(TextAndVersion.Create(SourceText.From(code),
+                                VersionStamp.Create())),
+                                filePath: fileName)
+                         })
+                });
+        }
+    }
+}
